Did folder name refactoring
