+++ lib/screens/buyer_signin.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:plug/components/constants.dart';
+import 'package:plug/utilities/constants.dart';
 
 class BuyerSignInPage extends StatelessWidget {
   const BuyerSignInPage({Key? key}) : super(key: key);
